Finished multiple components functionality
-------

3/10/16

Began Multiple Components tutorial
https://angular.io/docs/ts/latest/tutorial/toh-pt3.html

Due to the Single Responsibility Principle (modularity, per Uncle Bob), the different types of components will need to be broken out into separate files. 

"When you write a software module, you want to make sure that when changes are requested, those changes can only originate from a single person, or rather, a single tightly coupled group of people representing a single narrowly defined business function. You want to isolate your modules from the complexities of the organization as a whole, and design your systems such that each module is responsible (responds to) the needs of just that one business function.

This is the reason we do not put SQL in JSPs. This is the reason we do not generate HTML in the modules that compute results. This is the reason that business rules should not know the database schema. This is the reason we separate concer